# docs/swagger.yaml
basePath: /api
definitions:
...
  serializer.PrescriptionEditJSON:
    properties:
      doctor_full_name:
        type: string
      notes:
        type: string
    type: object
  serializer.PrescriptionJSON:
    properties:
      completed_dose_line_count:
        type: integer
      created_at:
        type: string
      creator_login:
        type: string
      doctor_full_name:
        type: string
      finish_date:
        type: string
      forming_date:
        type: string
      moderator_login:
        type: string
      notes:
        type: string
      prescription_id:
        type: integer
      status:
        type: string
    type: object
...
  serializer.StatusJSON:
    properties:
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://localhost:8080
  description: API для управления рецептами, препаратами и расчетом доз.
  license:
    name: MIT
  title: Prescription API
  version: "1.0"
paths:
...
  /prescriptions:
    get:
      description: Возвращает список рецептов с фильтрацией по дате и статусу.
      parameters:
      - description: Начальная дата (YYYY-MM-DD)
        in: query
        name: from-date
        type: string
      - description: Конечная дата (YYYY-MM-DD)
        in: query
        name: to-date
        type: string
      - description: Статус рецепта
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список рецептов
          schema:
            items:
              $ref: '#/definitions/serializer.PrescriptionJSON'
            type: array
        "400":
          description: Неверный формат даты
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список рецептов
      tags:
      - prescriptions
  /prescriptions/{id}:
    delete:
      description: Выполняет логическое удаление черновика рецепта.
      parameters:
      - description: ID рецепта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Рецепт удалён
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Рецепт не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить рецепт
      tags:
      - prescriptions
    get:
      description: Возвращает рецепт и список назначенных препаратов.
      parameters:
      - description: ID рецепта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Рецепт и препараты
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Рецепт не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить рецепт по ID
      tags:
      - prescriptions
    put:
      consumes:
      - application/json
      description: 'Обновляет поля черновика рецепта: ФИО врача и примечания.'
      parameters:
      - description: ID рецепта
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные рецепта
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.PrescriptionEditJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный рецепт
          schema:
            $ref: '#/definitions/serializer.PrescriptionJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Рецепт не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить рецепт
      tags:
      - prescriptions
  /prescriptions/{id}/finish:
    put:
      consumes:
      - application/json
      description: Меняет статус сформированного рецепта на completed или rejected
        (только модератор).
      parameters:
      - description: ID рецепта
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/serializer.StatusJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный рецепт
          schema:
            $ref: '#/definitions/serializer.PrescriptionJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Рецепт не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Завершить или отклонить рецепт
      tags:
      - prescriptions
  /prescriptions/{id}/form:
    put:
      description: Переводит рецепт из черновика в статус formed и пересчитывает дозы.
      parameters:
      - description: ID рецепта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сформированный рецепт
          schema:
            $ref: '#/definitions/serializer.PrescriptionJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Рецепт не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать рецепт
      tags:
      - prescriptions
  /prescriptions/cart:
    get:
      description: Возвращает информацию о текущем черновике рецепта пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: Данные корзины
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить корзину рецепта
      tags:
      - prescriptions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"